feat!: don't export findUppercaseKeys
