+++ test/unit/util/sfdcTest.ts
@@ -6,7 +6,6 @@
  */
 import { expect } from 'chai';
 import {
-  findUpperCaseKeys,
   matchesAccessToken,
   trimTo15,
   validateApiVersion,
@@ -100,60 +99,6 @@ describe('util/sfdc', () => {
     });
     it('should return true for "c:\\myfile"', () => {
       expect(validatePathDoesNotContainInvalidChars('c:\\myfile')).to.be.true;
-    });
-  });
-
-  describe('findUpperCaseKeys', () => {
-    it('should return the first upper case key', () => {
-      const testObj = {
-        lowercase: true,
-        UpperCase: false,
-        nested: { camelCase: true },
-      };
-      expect(findUpperCaseKeys(testObj)).to.equal('UpperCase');
-    });
-
-    it('should return the first nested upper case key', () => {
-      const testObj = {
-        lowercase: true,
-        uppercase: false,
-        nested: { NestedUpperCase: true },
-      };
-      expect(findUpperCaseKeys(testObj)).to.equal('NestedUpperCase');
-    });
-
-    it('should return undefined when no upper case key is found', () => {
-      const testObj = {
-        lowercase: true,
-        uppercase: false,
-        nested: { camelCase: true },
-      };
-      expect(findUpperCaseKeys(testObj)).to.be.undefined;
-    });
-
-    it('should return the first nested upper case key unless blocklisted', () => {
-      const testObj = {
-        lowercase: true,
-        uppercase: false,
-        nested: { NestedUpperCase: true },
-      };
-      expect(findUpperCaseKeys(testObj, ['nested'])).to.equal(undefined);
-    });
-
-    it('handles keys starting with numbers', () => {
-      const testObj = {
-        '1abc': true,
-        Abc: false,
-        nested: { '2abc': true },
-      };
-      expect(findUpperCaseKeys(testObj)).to.equal('Abc');
-    });
-    it('handles keys starting with numbers', () => {
-      const testObj = {
-        '1abc': true,
-        nested: { '2abc': true, Upper: false },
-      };
-      expect(findUpperCaseKeys(testObj)).to.equal('Upper');
     });
   });
